% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ files
\name{all_thetas_plotter}
\alias{all_thetas_plotter}
\title{all_thetas_plotter}
\usage{
all_thetas_plotter(credibility_model, N_groups, ages_fit, predictor)
}
\arguments{
\item{credibility_model}{Input used by \code{all_thetas_plotter}.}
\item{N_groups}{Input used by \code{all_thetas_plotter}.}
\item{ages_fit}{Input used by \code{all_thetas_plotter}.}
\item{predictor}{Input used by \code{all_thetas_plotter}.}
}
\value{
An R object produced by this step of the workflow.
}
\description{
Purpose: all_thetas_plotter for the mortality forecasting reproducibility workflow.
}
