% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ files
\name{average_smoother}
\alias{average_smoother}
\title{average_smoother}
\usage{
average_smoother(pos, x, k)
}
\arguments{
\item{pos}{Input used by \code{average_smoother}.}
\item{x}{Input used by \code{average_smoother}.}
\item{k}{Input used by \code{average_smoother}.}
}
\value{
An R object produced by this step of the workflow.
}
\description{
Purpose: average_smoother for the mortality forecasting reproducibility workflow.
}
