% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ files
\name{bind_separate_failures}
\alias{bind_separate_failures}
\title{bind_separate_failures}
\usage{
bind_separate_failures(x)
}
\arguments{
\item{x}{Input used by \code{bind_separate_failures}.}
}
\value{
An R object produced by this step of the workflow.
}
\description{
Purpose: bind_separate_failures for the mortality forecasting reproducibility workflow.
}
