% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ files
\name{binomial_simulator}
\alias{binomial_simulator}
\title{binomial_simulator}
\usage{
binomial_simulator(starting_exposure, probabilities)
}
\arguments{
\item{starting_exposure}{Input used by \code{binomial_simulator}.}
\item{probabilities}{Input used by \code{binomial_simulator}.}
}
\value{
An R object produced by this step of the workflow.
}
\description{
Purpose: binomial_simulator for the mortality forecasting reproducibility workflow.
}
