% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generators.R
\name{data_generator_hmd_lt}
\alias{data_generator_hmd_lt}
\title{data_generator_hmd_lt}
\usage{
data_generator_hmd_lt(
  seed_input = 1,
  hmd_username = NULL,
  hmd_password = NULL,
  exposure_sup = 1e+05,
  exposure_sub = c(5000, 500)
)
}
\arguments{
\item{seed_input}{Input used by \code{data_generator_hmd_lt}.}

\item{hmd_username}{Input used by \code{data_generator_hmd_lt}.}

\item{hmd_password}{Input used by \code{data_generator_hmd_lt}.}

\item{exposure_sup}{Input used by \code{data_generator_hmd_lt}.}

\item{exposure_sub}{Input used by \code{data_generator_hmd_lt}.}
}
\value{
An R object produced by this step of the workflow.
}
\description{
Purpose: data_generator_hmd_lt for the mortality forecasting reproducibility workflow.
}
