% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{data_preprocessing}
\alias{data_preprocessing}
\title{data_preprocessing}
\usage{
data_preprocessing(
  data,
  N_groups = 3,
  bias = 0,
  years_fit = 110:150,
  ages_fit = 50:65
)
}
\arguments{
\item{data}{Input used by \code{data_preprocessing}.}

\item{N_groups}{Input used by \code{data_preprocessing}.}

\item{bias}{Input used by \code{data_preprocessing}.}

\item{years_fit}{Input used by \code{data_preprocessing}.}

\item{ages_fit}{Input used by \code{data_preprocessing}.}
}
\value{
An R object produced by this step of the workflow.
}
\description{
Purpose: data_preprocessing for the mortality forecasting reproducibility workflow.
}
