% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ files
\name{fit_and_predict_credibility_models}
\alias{fit_and_predict_credibility_models}
\title{fit_and_predict_credibility_models}
\usage{
fit_and_predict_credibility_models(data, N_groups, mortality_models_fit, years_fit, ages_fit, bias, forecasting_horizon)
}
\arguments{
\item{data}{Input used by \code{fit_and_predict_credibility_models}.}
\item{N_groups}{Input used by \code{fit_and_predict_credibility_models}.}
\item{mortality_models_fit}{Input used by \code{fit_and_predict_credibility_models}.}
\item{years_fit}{Input used by \code{fit_and_predict_credibility_models}.}
\item{ages_fit}{Input used by \code{fit_and_predict_credibility_models}.}
\item{bias}{Input used by \code{fit_and_predict_credibility_models}.}
\item{forecasting_horizon}{Input used by \code{fit_and_predict_credibility_models}.}
}
\value{
An R object produced by this step of the workflow.
}
\description{
Purpose: fit_and_predict_credibility_models for the mortality forecasting reproducibility workflow.
}
