% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_credibility.R
\name{fit_and_predict_separate_models}
\alias{fit_and_predict_separate_models}
\title{fit_and_predict_separate_models}
\usage{
fit_and_predict_separate_models(
  data,
  N_groups,
  mortality_models_fit,
  years_fit,
  ages_fit,
  bias = 70,
  forecasting_horizon,
  max_ratio_implausible = 50,
  max_log_jump_implausible = 3.5
)
}
\arguments{
\item{data}{Input used by \code{fit_and_predict_separate_models}.}

\item{N_groups}{Input used by \code{fit_and_predict_separate_models}.}

\item{mortality_models_fit}{Input used by \code{fit_and_predict_separate_models}.}

\item{years_fit}{Input used by \code{fit_and_predict_separate_models}.}

\item{ages_fit}{Input used by \code{fit_and_predict_separate_models}.}

\item{bias}{Input used by \code{fit_and_predict_separate_models}.}

\item{forecasting_horizon}{Input used by \code{fit_and_predict_separate_models}.}

\item{max_ratio_implausible}{Input used by \code{fit_and_predict_separate_models}.}

\item{max_log_jump_implausible}{Input used by \code{fit_and_predict_separate_models}.}
}
\value{
An R object produced by this step of the workflow.
}
\description{
Purpose: fit_and_predict_separate_models for the mortality forecasting reproducibility workflow.
}
