% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_credibility.R
\name{fit_and_predict_total_model}
\alias{fit_and_predict_total_model}
\title{fit_and_predict_total_model}
\usage{
fit_and_predict_total_model(
  data,
  N_groups,
  mortality_models_fit,
  years_fit,
  ages_fit,
  bias = 70,
  forecasting_horizon
)
}
\arguments{
\item{data}{Input used by \code{fit_and_predict_total_model}.}

\item{N_groups}{Input used by \code{fit_and_predict_total_model}.}

\item{mortality_models_fit}{Input used by \code{fit_and_predict_total_model}.}

\item{years_fit}{Input used by \code{fit_and_predict_total_model}.}

\item{ages_fit}{Input used by \code{fit_and_predict_total_model}.}

\item{bias}{Input used by \code{fit_and_predict_total_model}.}

\item{forecasting_horizon}{Input used by \code{fit_and_predict_total_model}.}
}
\value{
An R object produced by this step of the workflow.
}
\description{
Purpose: fit_and_predict_total_model for the mortality forecasting reproducibility workflow.
}
