% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_credibility.R
\name{fit_mortality_models}
\alias{fit_mortality_models}
\title{fit_mortality_models}
\usage{
fit_mortality_models(
  data_pp,
  years_fit,
  ages_fit,
  global_mortality_model_candidates = c("lc", "apc", "rh")
)
}
\arguments{
\item{data_pp}{Input used by \code{fit_mortality_models}.}

\item{years_fit}{Input used by \code{fit_mortality_models}.}

\item{ages_fit}{Input used by \code{fit_mortality_models}.}

\item{global_mortality_model_candidates}{Input used by \code{fit_mortality_models}.}
}
\value{
An R object produced by this step of the workflow.
}
\description{
Purpose: fit_mortality_models for the mortality forecasting reproducibility workflow.
}
