% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_credibility.R
\name{model_assessment_full_mle}
\alias{model_assessment_full_mle}
\title{model_assessment_full_mle}
\usage{
model_assessment_full_mle(
  model_fit_and_prediction,
  N_groups,
  years_fit,
  forecasting_horizon,
  bias = 15,
  subset_rates = NULL
)
}
\arguments{
\item{model_fit_and_prediction}{Input used by \code{model_assessment_full_mle}.}

\item{N_groups}{Input used by \code{model_assessment_full_mle}.}

\item{years_fit}{Input used by \code{model_assessment_full_mle}.}

\item{forecasting_horizon}{Input used by \code{model_assessment_full_mle}.}

\item{bias}{Input used by \code{model_assessment_full_mle}.}

\item{subset_rates}{Input used by \code{model_assessment_full_mle}.}
}
\value{
An R object produced by this step of the workflow.
}
\description{
Purpose: model_assessment_full_mle for the mortality forecasting reproducibility workflow.
}
