% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ files
\name{model_assessment_superpop}
\alias{model_assessment_superpop}
\title{model_assessment_superpop}
\usage{
model_assessment_superpop(model_fit_and_prediction, N_groups, years_fit, forecasting_horizon, bias, subset_rates)
}
\arguments{
\item{model_fit_and_prediction}{Input used by \code{model_assessment_superpop}.}
\item{N_groups}{Input used by \code{model_assessment_superpop}.}
\item{years_fit}{Input used by \code{model_assessment_superpop}.}
\item{forecasting_horizon}{Input used by \code{model_assessment_superpop}.}
\item{bias}{Input used by \code{model_assessment_superpop}.}
\item{subset_rates}{Input used by \code{model_assessment_superpop}.}
}
\value{
An R object produced by this step of the workflow.
}
\description{
Purpose: model_assessment_superpop for the mortality forecasting reproducibility workflow.
}
