% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ files
\name{plot_msep_group}
\alias{plot_msep_group}
\title{plot_msep_group}
\usage{
plot_msep_group(group_result, ylim)
}
\arguments{
\item{group_result}{Input used by \code{plot_msep_group}.}
\item{ylim}{Input used by \code{plot_msep_group}.}
}
\value{
An R object produced by this step of the workflow.
}
\description{
Purpose: plot_msep_group for the mortality forecasting reproducibility workflow.
}
