% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ files
\name{plot_predicted_event_trends}
\alias{plot_predicted_event_trends}
\title{plot_predicted_event_trends}
\usage{
plot_predicted_event_trends(mortality_models_fit, data_pp, # subgroup, model_option)
}
\arguments{
\item{mortality_models_fit}{Input used by \code{plot_predicted_event_trends}.}
\item{data_pp}{Input used by \code{plot_predicted_event_trends}.}
\item{# subgroup}{Input used by \code{plot_predicted_event_trends}.}
\item{model_option}{Input used by \code{plot_predicted_event_trends}.}
}
\value{
An R object produced by this step of the workflow.
}
\description{
Purpose: plot_predicted_event_trends for the mortality forecasting reproducibility workflow.
}
