% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ files
\name{print_age_effects}
\alias{print_age_effects}
\title{print_age_effects}
\usage{
print_age_effects(seed_input, hmd_username, hmd_password, exposure_sup, exposure_sub)
}
\arguments{
\item{seed_input}{Input used by \code{print_age_effects}.}
\item{hmd_username}{Input used by \code{print_age_effects}.}
\item{hmd_password}{Input used by \code{print_age_effects}.}
\item{exposure_sup}{Input used by \code{print_age_effects}.}
\item{exposure_sub}{Input used by \code{print_age_effects}.}
}
\value{
An R object produced by this step of the workflow.
}
\description{
Purpose: print_age_effects for the mortality forecasting reproducibility workflow.
}
