% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ files
\name{run_msep_study}
\alias{run_msep_study}
\title{run_msep_study}
\usage{
run_msep_study(data, data_pp, mortality_models_fit, credibility_model, ages_fit, years_fit, N_groups, predictor, reference_age, forecasting_horizon, bias)
}
\arguments{
\item{data}{Input used by \code{run_msep_study}.}
\item{data_pp}{Input used by \code{run_msep_study}.}
\item{mortality_models_fit}{Input used by \code{run_msep_study}.}
\item{credibility_model}{Input used by \code{run_msep_study}.}
\item{ages_fit}{Input used by \code{run_msep_study}.}
\item{years_fit}{Input used by \code{run_msep_study}.}
\item{N_groups}{Input used by \code{run_msep_study}.}
\item{predictor}{Input used by \code{run_msep_study}.}
\item{reference_age}{Input used by \code{run_msep_study}.}
\item{forecasting_horizon}{Input used by \code{run_msep_study}.}
\item{bias}{Input used by \code{run_msep_study}.}
}
\value{
An R object produced by this step of the workflow.
}
\description{
Purpose: run_msep_study for the mortality forecasting reproducibility workflow.
}
