% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_credibility.R
\name{sample_size_effect_on_predictions}
\alias{sample_size_effect_on_predictions}
\title{sample_size_effect_on_predictions}
\usage{
sample_size_effect_on_predictions(
  data,
  years_fit_basic,
  number_of_rolls,
  ages_fit,
  N_groups = 3,
  bias = 70,
  global_mortality_model = "lc"
)
}
\arguments{
\item{data}{Input used by \code{sample_size_effect_on_predictions}.}

\item{years_fit_basic}{Input used by \code{sample_size_effect_on_predictions}.}

\item{number_of_rolls}{Input used by \code{sample_size_effect_on_predictions}.}

\item{ages_fit}{Input used by \code{sample_size_effect_on_predictions}.}

\item{N_groups}{Input used by \code{sample_size_effect_on_predictions}.}

\item{bias}{Input used by \code{sample_size_effect_on_predictions}.}

\item{global_mortality_model}{Input used by \code{sample_size_effect_on_predictions}.}
}
\value{
An R object produced by this step of the workflow.
}
\description{
Purpose: sample_size_effect_on_predictions for the mortality forecasting reproducibility workflow.
}
