% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_credibility.R
\name{sample_size_effect_plotter}
\alias{sample_size_effect_plotter}
\title{sample_size_effect_plotter}
\usage{
sample_size_effect_plotter(ss_effect, color_group, ages_to_plot, label_to_plot)
}
\arguments{
\item{ss_effect}{Input used by \code{sample_size_effect_plotter}.}

\item{color_group}{Input used by \code{sample_size_effect_plotter}.}

\item{ages_to_plot}{Input used by \code{sample_size_effect_plotter}.}

\item{label_to_plot}{Input used by \code{sample_size_effect_plotter}.}
}
\value{
An R object produced by this step of the workflow.
}
\description{
Purpose: sample_size_effect_plotter for the mortality forecasting reproducibility workflow.
}
