% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_credibility.R
\name{scoring_plot}
\alias{scoring_plot}
\title{scoring_plot}
\usage{
scoring_plot(
  data,
  metric,
  superpopulation = FALSE,
  subpopulation = FALSE,
  predictor_choice
)
}
\arguments{
\item{data}{Input used by \code{scoring_plot}.}

\item{metric}{Input used by \code{scoring_plot}.}

\item{superpopulation}{Input used by \code{scoring_plot}.}

\item{subpopulation}{Input used by \code{scoring_plot}.}

\item{predictor_choice}{Input used by \code{scoring_plot}.}
}
\value{
An R object produced by this step of the workflow.
}
\description{
Purpose: scoring_plot for the mortality forecasting reproducibility workflow.
}
