% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ files
\name{thetas_plotter}
\alias{thetas_plotter}
\title{thetas_plotter}
\usage{
thetas_plotter(credibility_model, subgroup, ages_fit, years_fit, mortality_models_fit, predictor, chosen_age)
}
\arguments{
\item{credibility_model}{Input used by \code{thetas_plotter}.}
\item{subgroup}{Input used by \code{thetas_plotter}.}
\item{ages_fit}{Input used by \code{thetas_plotter}.}
\item{years_fit}{Input used by \code{thetas_plotter}.}
\item{mortality_models_fit}{Input used by \code{thetas_plotter}.}
\item{predictor}{Input used by \code{thetas_plotter}.}
\item{chosen_age}{Input used by \code{thetas_plotter}.}
}
\value{
An R object produced by this step of the workflow.
}
\description{
Purpose: thetas_plotter for the mortality forecasting reproducibility workflow.
}
