% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_credibility.R
\name{variance_plotter}
\alias{variance_plotter}
\title{variance_plotter}
\usage{
variance_plotter(credibility_model, predictor = "lc", ages_fit)
}
\arguments{
\item{credibility_model}{Input used by \code{variance_plotter}.}

\item{predictor}{Input used by \code{variance_plotter}.}

\item{ages_fit}{Input used by \code{variance_plotter}.}
}
\value{
An R object produced by this step of the workflow.
}
\description{
Purpose: variance_plotter for the mortality forecasting reproducibility workflow.
}
