% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ files
\name{weights_plotter}
\alias{weights_plotter}
\title{weights_plotter}
\usage{
weights_plotter(credibility_model, N_groups, ages_fit, predictor, ages_breaks)
}
\arguments{
\item{credibility_model}{Input used by \code{weights_plotter}.}
\item{N_groups}{Input used by \code{weights_plotter}.}
\item{ages_fit}{Input used by \code{weights_plotter}.}
\item{predictor}{Input used by \code{weights_plotter}.}
\item{ages_breaks}{Input used by \code{weights_plotter}.}
}
\value{
An R object produced by this step of the workflow.
}
\description{
Purpose: weights_plotter for the mortality forecasting reproducibility workflow.
}
