% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_credibility.R
\name{write_msep_plot}
\alias{write_msep_plot}
\title{write_msep_plot}
\usage{
write_msep_plot(group_result, out, ylim = NULL)
}
\arguments{
\item{group_result}{Input used by \code{write_msep_plot}.}

\item{out}{Input used by \code{write_msep_plot}.}

\item{ylim}{Input used by \code{write_msep_plot}.}
}
\value{
An R object produced by this step of the workflow.
}
\description{
Purpose: write_msep_plot for the mortality forecasting reproducibility workflow.
}
