% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ files
\name{data_generator_hmd_lt}
\alias{data_generator_hmd_lt}
\title{data_generator_hmd_lt}
\usage{
data_generator_hmd_lt(seed_input, hmd_username, hmd_password, exposure_sup, exposure_sub)
}
\arguments{
\item{seed_input}{Input used by \code{data_generator_hmd_lt}.}
\item{hmd_username}{Input used by \code{data_generator_hmd_lt}.}
\item{hmd_password}{Input used by \code{data_generator_hmd_lt}.}
\item{exposure_sup}{Input used by \code{data_generator_hmd_lt}.}
\item{exposure_sub}{Input used by \code{data_generator_hmd_lt}.}
}
\value{
An R object produced by this step of the workflow.
}
\description{
Purpose: data_generator_hmd_lt for the mortality forecasting reproducibility workflow.
}
