% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_credibility.R
\name{k_years_smoothing}
\alias{k_years_smoothing}
\title{k_years_smoothing}
\usage{
k_years_smoothing(x, k = 5)
}
\arguments{
\item{x}{Input used by \code{k_years_smoothing}.}

\item{k}{Input used by \code{k_years_smoothing}.}
}
\value{
An R object produced by this step of the workflow.
}
\description{
Purpose: k_years_smoothing for the mortality forecasting reproducibility workflow.
}
