% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ files
\name{model_assessment}
\alias{model_assessment}
\title{model_assessment}
\usage{
model_assessment(model_fit_and_prediction, N_groups, years_fit, forecasting_horizon, bias, subset_rates)
}
\arguments{
\item{model_fit_and_prediction}{Input used by \code{model_assessment}.}
\item{N_groups}{Input used by \code{model_assessment}.}
\item{years_fit}{Input used by \code{model_assessment}.}
\item{forecasting_horizon}{Input used by \code{model_assessment}.}
\item{bias}{Input used by \code{model_assessment}.}
\item{subset_rates}{Input used by \code{model_assessment}.}
}
\value{
An R object produced by this step of the workflow.
}
\description{
Updated model assessment compatible with both old and new fit_and_predict_* outputs.
}
