% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ files
\name{plot_credibility_mle_global}
\alias{plot_credibility_mle_global}
\title{plot_credibility_mle_global}
\usage{
plot_credibility_mle_global(credibility_model, N_groups, ages_fit, predictor, ages_breaks)
}
\arguments{
\item{credibility_model}{Input used by \code{plot_credibility_mle_global}.}
\item{N_groups}{Input used by \code{plot_credibility_mle_global}.}
\item{ages_fit}{Input used by \code{plot_credibility_mle_global}.}
\item{predictor}{Input used by \code{plot_credibility_mle_global}.}
\item{ages_breaks}{Input used by \code{plot_credibility_mle_global}.}
}
\value{
An R object produced by this step of the workflow.
}
\description{
Purpose: plot_credibility_mle_global for the mortality forecasting reproducibility workflow.
}
