% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ files
\name{plot_smoothed_unsmoothed}
\alias{plot_smoothed_unsmoothed}
\title{plot_smoothed_unsmoothed}
\usage{
plot_smoothed_unsmoothed(data, data_pp, N_groups, mortality_models_fit, years_fit, ages_fit, bias, scenario, model_2_plot, smoothing_years, forecasting_horizon)
}
\arguments{
\item{data}{Input used by \code{plot_smoothed_unsmoothed}.}
\item{data_pp}{Input used by \code{plot_smoothed_unsmoothed}.}
\item{N_groups}{Input used by \code{plot_smoothed_unsmoothed}.}
\item{mortality_models_fit}{Input used by \code{plot_smoothed_unsmoothed}.}
\item{years_fit}{Input used by \code{plot_smoothed_unsmoothed}.}
\item{ages_fit}{Input used by \code{plot_smoothed_unsmoothed}.}
\item{bias}{Input used by \code{plot_smoothed_unsmoothed}.}
\item{scenario}{Input used by \code{plot_smoothed_unsmoothed}.}
\item{model_2_plot}{Input used by \code{plot_smoothed_unsmoothed}.}
\item{smoothing_years}{Input used by \code{plot_smoothed_unsmoothed}.}
\item{forecasting_horizon}{Input used by \code{plot_smoothed_unsmoothed}.}
}
\value{
An R object produced by this step of the workflow.
}
\description{
Purpose: plot_smoothed_unsmoothed for the mortality forecasting reproducibility workflow.
}
