% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generators.R
\name{print_age_effects}
\alias{print_age_effects}
\title{print_age_effects}
\usage{
print_age_effects(
  seed_input = 1,
  hmd_username = NULL,
  hmd_password = NULL,
  exposure_sup = 1e+05,
  exposure_sub = c(5000, 500)
)
}
\arguments{
\item{seed_input}{Input used by \code{print_age_effects}.}

\item{hmd_username}{Input used by \code{print_age_effects}.}

\item{hmd_password}{Input used by \code{print_age_effects}.}

\item{exposure_sup}{Input used by \code{print_age_effects}.}

\item{exposure_sub}{Input used by \code{print_age_effects}.}
}
\value{
An R object produced by this step of the workflow.
}
\description{
Purpose: print_age_effects for the mortality forecasting reproducibility workflow.
}
