% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ files
\name{run_insample_bic_cell}
\alias{run_insample_bic_cell}
\title{run_insample_bic_cell}
\usage{
run_insample_bic_cell(seed_ix, years_ix, ages_fit, years_fit_basic, N_groups, exposure_sup, exposure_sub)
}
\arguments{
\item{seed_ix}{Input used by \code{run_insample_bic_cell}.}
\item{years_ix}{Input used by \code{run_insample_bic_cell}.}
\item{ages_fit}{Input used by \code{run_insample_bic_cell}.}
\item{years_fit_basic}{Input used by \code{run_insample_bic_cell}.}
\item{N_groups}{Input used by \code{run_insample_bic_cell}.}
\item{exposure_sup}{Input used by \code{run_insample_bic_cell}.}
\item{exposure_sub}{Input used by \code{run_insample_bic_cell}.}
}
\value{
An R object produced by this step of the workflow.
}
\description{
Purpose: run_insample_bic_cell for the mortality forecasting reproducibility workflow.
}
