% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_credibility.R
\name{run_separate_failures_cell}
\alias{run_separate_failures_cell}
\title{run_separate_failures_cell}
\usage{
run_separate_failures_cell(
  seed_ix,
  years_ix,
  ages_fit = 15:85,
  years_fit_basic = 118:140,
  N_groups = 3,
  forecasting_horizon = 1,
  exposure_sup = 94500,
  exposure_sub = c(5000, 500),
  global_mortality_model = c("lc", "apc", "rh"),
  hmd_user = "",
  hmd_pass = ""
)
}
\value{
An R object produced by this step of the workflow.
}
\description{
Purpose: run_separate_failures_cell for the mortality forecasting reproducibility workflow.
}
